Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ultralytics>=8.0.196
pillow>=10.1.0
python-multipart==0.0.6
torch>=2.0.0
numpy>=1.26.2
opencv-python>=4.8.1.78
pydantic>=2.0.0 